+++ packages/qwik/src/utils/index.ts
@@ -25,6 +25,7 @@ export * from './setFieldArrayState';
 export * from './setFieldArrayValue';
 export * from './setFieldErrors';
 export * from './setFieldState';
+export * from './sortArrayPathIndex';
 export * from './updateFieldArrayDirty';
 export * from './updateFieldDirty';
 export * from './updateFormDirty';