+++ packages/react/src/methods/insert.ts
@@ -2,7 +2,6 @@ import { batch } from '@preact/signals-react';
 import type {
   FieldArrayPath,
   FieldArrayPathValue,
-  FieldPath,
   FieldValues,
   FormStore,
   Maybe,
@@ -17,6 +16,7 @@ import {
   setFieldArrayState,
   setFieldArrayValue,
   setFieldState,
+  sortArrayPathIndex,
   validateIfRequired,
 } from '../utils';
 
@@ -68,12 +68,6 @@ export function insert<
           const filterName = (value: string) =>
             value.startsWith(`${name}.`) && getPathIndex(name, value) >= index;
 
-          // Create function to sort by path index
-          const sortPathIndex = (
-            pathA: FieldPath<TFieldValues> | FieldArrayPath<TFieldValues>,
-            pathB: FieldPath<TFieldValues> | FieldArrayPath<TFieldValues>
-          ) => getPathIndex(name, pathA) - getPathIndex(name, pathB);
-
           // Create function to get next index name
           const getNextIndexName = <T extends string>(
             fieldOrFieldArrayName: T,
@@ -88,7 +82,7 @@ export function insert<
           form.internal.fieldNames
             .peek()
             .filter(filterName)
-            .sort(sortPathIndex)
+            .sort(sortArrayPathIndex(name))
             .reverse()
             .forEach((fieldName) => {
               setFieldState(
@@ -102,7 +96,7 @@ export function insert<
           form.internal.fieldArrayNames
             .peek()
             .filter(filterName)
-            .sort(sortPathIndex)
+            .sort(sortArrayPathIndex(name))
             .reverse()
             .forEach((fieldArrayName) => {
               setFieldArrayState(