+++ packages/solid/src/methods/move.ts
@@ -15,6 +15,7 @@ import {
   getPathIndex,
   setFieldArrayState,
   setFieldState,
+  sortArrayPathIndex,
   updateFieldArrayDirty,
 } from '../utils';
 
@@ -70,12 +71,6 @@ export function move<
           }
         };
 
-        // Create function to sort by path index
-        const sortPathIndex = (
-          pathA: FieldPath<TFieldValues> | FieldArrayPath<TFieldValues>,
-          pathB: FieldPath<TFieldValues> | FieldArrayPath<TFieldValues>
-        ) => getPathIndex(name, pathA) - getPathIndex(name, pathB);
-
         // Create function to get previous index name
         const getPrevIndexName = <T extends string>(
           fieldOrFieldArrayName: T,
@@ -101,11 +96,11 @@ export function move<
         const fieldNames = form.internal.fieldNames
           .get()
           .filter(filterName)
-          .sort(sortPathIndex);
+          .sort(sortArrayPathIndex(name));
         const fieldArrayNames = form.internal.fieldArrayNames
           .get()
           .filter(filterName)
-          .sort(sortPathIndex);
+          .sort(sortArrayPathIndex(name));
 
         // Reverse names if "from" index is greater than "to" index
         if (fromIndex > toIndex) {