+++ packages/solid/src/methods/remove.ts
@@ -1,7 +1,6 @@
 import { batch, untrack } from 'solid-js';
 import type {
   FieldArrayPath,
-  FieldPath,
   FieldValues,
   FormStore,
   ResponseData,
@@ -13,6 +12,7 @@ import {
   getPathIndex,
   setFieldArrayState,
   setFieldState,
+  sortArrayPathIndex,
   updateFieldArrayDirty,
   validateIfRequired,
 } from '../utils';
@@ -54,12 +54,6 @@ export function remove<
         const filterName = (value: string) =>
           value.startsWith(`${name}.`) && getPathIndex(name, value) > index;
 
-        // Create function to sort by path index
-        const sortPathIndex = (
-          pathA: FieldPath<TFieldValues> | FieldArrayPath<TFieldValues>,
-          pathB: FieldPath<TFieldValues> | FieldArrayPath<TFieldValues>
-        ) => getPathIndex(name, pathA) - getPathIndex(name, pathB);
-
         // Create function to get previous index name
         const getPrevIndexName = <T extends string>(
           fieldOrFieldArrayName: T,
@@ -75,7 +69,7 @@ export function remove<
           form.internal.fieldNames
             .get()
             .filter(filterName)
-            .sort(sortPathIndex)
+            .sort(sortArrayPathIndex(name))
             .forEach((fieldName) => {
               setFieldState(
                 form,
@@ -88,7 +82,7 @@ export function remove<
           form.internal.fieldArrayNames
             .get()
             .filter(filterName)
-            .sort(sortPathIndex)
+            .sort(sortArrayPathIndex(name))
             .forEach((fieldArrayName) => {
               setFieldArrayState(
                 form,